ai-doc-qa-backend/requirements.txt
fastapi
uvicorn

llama-index-core
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-vector-stores-faiss

faiss-cpu
pypdf
python-multipart
sentence-transformers
